Add `bubble_sort` method to sort the array
 + Sort is calculated with two nested loops.
 + Inner loop is to iterate over each element.
 + Outer loop is to iterate the entire array (or) to complete one cycle of sort where the last element is sorted.

Optimizations are done to reduce the inner loop after each cycle / outer loop iteration.
As the last element is already in place, no. of inner loop is set to be reduced.
If there are no modifications done, it is set to break out of loop and return the "sorted array".