Add `class instance variables` & `class methods`
 + Added class instance variable -> "@available_positions" along with getter & setter method as "attr_accessor".
 + Any instance of this class can able to access & modify this variable.
 + Additionally, "valid_position?(number)" class method is added to operate on the class level.

Due to the requirement of shared state, a "class instance variable" and "class methods" are used.
It's use case comes handy in "Edge case scenario -> Choosing the already selected position".
If this method is unavailable, the place where 'X' is present can be replaces with 'O' by choosing same position.
This method helps in filtering out the positions by removing the used positions from the class instance array.