Fix `HashMap#clear` method
 + If "clear()" is invoked, it now creates a new "bucket_stash" & also gets new "bucket".

Before Change ==> It just re-assigns all the values to "nil".
But, if the bucket size is more than 100, all 100 would be re-assigned, which is waste of space.
After Change ==> It now creates a new "bucket_stash" from "Buckets" class.
Now, if the bucket size if more than 100, it will change back to default size "16" with "nil" values.